Add an awake check for OnValidate
When loading a prefab OnValidate will get called before the MonoBehaviour is loaded, this causes IsPartOfPrefabAsset to load the whole file, causing SendMessage to get called from MeshFilter.
By checking in OnValidate to see if we are an Awake instance of an object, the case we care about here, we can prevent the erroneous warning.